Revert "Merge branch 'develop' into origin/feature/PAL-11"
This reverts commit 456f85880e1c74e16ca1532a672d813d677645e1, reversing
changes made to 8cba7bbf862236daeafaa2246111f488b4a26c25.

+++ app/src/main/java/com/kosti/palesoccerfieldadmin/userListPackage/UserList.kt
@@ -70,7 +70,7 @@ class UserList : AppCompatActivity(), ProfileScreen.OnDismissListener {
                 data.putString("classification", userList[p2].Clasification)
                 data.putString("nickname", userList[p2].Nickname)
                 // Calculate age from Timestamp
-                data.putString("age", ((Date().time - userList[p2].Age.toLong()).toString()))
+                data.putString("age", ((Date().time - userList[p2].Age.toInt()).toString()))
                 data.putString("phone", userList[p2].Phone)
                 data.putStringArrayList("positions", userList[p2].Positions as ArrayList<String>)
                 data.putString("id", userList[p2].Id)
